inserting of entities with existing associated entities (not new associated entities) validation/whitelist of post url code refactor tests
